pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kn0"
version = "0.1.0"
description = "Entity & Relationship Extraction Engine — Knowledge Graph Builder"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "spacy>=3.7.0",
    "pymupdf>=1.24.0",
    "typer>=0.12.0",
    "pydantic-settings>=2.3.0",
    "networkx>=3.3.0",
    "pyvis>=0.3.2",
    "python-multipart>=0.0.9",
    "python-magic>=0.4.27",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "python-docx>=1.1.0",
    "openai>=1.30.0",
    "jinja2>=3.1.0",
]

[project.scripts]
kn0 = "kn0.cli:app"

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.25.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-beautifulsoup4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kn0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
